+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## v0.10.0-B1910036 (pre-release)
+
 - Added dependency `DependsOn` information to results from `Get-PSRule`. [#210](https://github.com/BernieWhite/PSRule/issues/210)
   - To include dependencies that would normally be filtered out use `-IncludeDependencies`.
 - Added input format for reading markdown front matter. [#301](https://github.com/BernieWhite/PSRule/issues/301)